Add function shared across tests in a header
