+++ test/CMakeLists.txt
@@ -29,7 +29,11 @@ set(SRC
 	./cpp/csr.cpp
     ./cpp/cg.cpp
 )
-add_executable(${PROJECT_NAME} ${SRC})
+
+SET(HEADERS
+    include/test_common.h
+)
+add_executable(${PROJECT_NAME} ${SRC} ${HEADERS})
 target_link_libraries(${PROJECT_NAME} doctest::doctest sparse_matrix_math)
 target_compile_definitions(${PROJECT_NAME} PRIVATE ASSET_PATH="${CMAKE_CURRENT_SOURCE_DIR}/assets/")
 target_include_directories(${PROJECT_NAME} PRIVATE include)
